refactor(test): replace sleep-heavy waits with shared timing (#24)
Unify the timing-sensitive parts of the suite around explicit wait helpers and coordination primitives so worker, collector, and adversarial coverage stop depending on fixed sleeps and manual polling.

test/common:
- expand `test/common/timing_optimization.rs` with shared polling and state helpers such as `wait_for_worker_status`, `wait_for_worker_processed_events`, and `wait_for_work_queue_status_count`, giving tests one place to wait on observable queue and worker state.
- add streamlining and coverage-assurance support modules, then extract helper code out of ad hoc test files so integration, system, and adversarial suites share the same synchronization surface.
- replace short scheduling sleeps with `tokio::task::yield_now()` where a test only needed cooperative progress, which keeps fast paths fast without stretching the wall-clock runtime.

test suites:
- rewrite worker, lifecycle, pipeline, race, backpressure, failure-mode, and stress scenarios to use barriers, queue-depth waits, and heartbeat-driven orphan detection instead of hand-rolled loops that could hang or overshoot.
- update `test/integration/collector/multi_source_coordination_test.rs`, `test/integration/failure_modes/*`, and `test/system/performance/*` to wait on queue state and processed-event counts, so assertions line up with real system progress instead of guessed delays.
- lower oversized adversarial timeout thresholds now that readiness and queue helpers exist, making regressions fail sooner and stopping long sleeps from masking deadlocks.

sinex-core / sources:
- add `crate/sinex-core/src/event_source_base.rs` with `EventSourceBase`, `parse_config`, `create_event`, `send_event`, and the `impl_event_source!` macro so event sources share config parsing and event construction code.
- adopt `EventSourceBase` across clipboard, filesystem, and Kitty setup paths, removing repeated config decoding and raw event assembly that the rewritten tests kept brushing against.

sinex-db / runtime:
- switch collector and worker code onto `RawEventBuilder` plus shared query helpers, so runtime code and test helpers stop maintaining separate event-construction paths.
- re-export common event and work-queue queries from `crate/sinex-db/src/lib.rs`, including `refresh_routing_cache`, `run_batch_router`, and `calculate_queue_depth_metrics`, so tests and runtime callers use the same surfaces instead of local copies.

sinex-ulid:
- add the `arbitrary` feature and `Arbitrary for Ulid` support, plus `arb_ulid`, `arb_timestamp`, `arb_hostname`, `arb_version`, and `arb_json_value`, so streamlined property tests stop carrying custom ULID scaffolding.

+++ test/property/mod.rs
@@ -26,7 +26,7 @@ pub mod strategies {
     /// Strategy for generating valid ULID timestamps
     pub fn valid_timestamps() -> impl Strategy<Value = DateTime<Utc>> {
         (0u64..2_000_000_000u64)  // Valid Unix timestamp range
-            .prop_map(|ts| DateTime::from_timestamp(ts, 0).unwrap_or(Utc::now()))
+            .prop_map(|ts| DateTime::from_timestamp(ts as i64, 0).unwrap_or(Utc::now()))
     }
 
     /// Strategy for generating realistic event payloads
